fix: Add common used methods into the utils folder
